complete chapter 8-9 implements skeletons with react suspense.
